# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dead Man's Slot Machine API
  description: Wild West themed slot machine game API (simulation only, no real money)
  version: 0.5.0
  contact:
    name: Tin Minarik
    email: tinminarik00@gmail.com
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Create a new account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password]
              properties:
                email:
                  type: string
                  format: email
                  example: cowboy@deadmans.com
                username:
                  type: string
                  minLength: 3
                  maxLength: 20
                  example: wild_bill
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  example: shootFirst123
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  accessToken:
                    type: string
        "409":
          description: Email or username already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: cowboy@deadmans.com
                password:
                  type: string
                  example: shootFirst123
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Get a new access token using a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        "401":
          description: Invalid or expired refresh token
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out (revoke refresh token)
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      description: Always returns success to prevent account enumeration
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: cowboy@deadmans.com
      responses:
        "200":
          description: Reset email sent (if account exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using a token from email
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Invalid or expired reset token
...
  /me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...